Create helper mixin
 ## Summary
These two test classes share a number of util functions; these can be
split out into a mixin for Test classes so it's easier to keep this code
in sync.
