+++ tests/helpers.py
@@ -330,6 +330,126 @@ class FixtureMixin(object):
         return search.id
 
 
+class PutDeclarationAndDetailsAndServicesMixin:
+    """Centralised from TestUsersExport and TestSuppliersExport; expects some attributes in the class this is mixed
+    into (self.supplier_id, self.framework_slug, self.client, self.updater_json)."""
+
+    def _register_supplier_with_framework(self):
+        response = self.client.put(
+            '/suppliers/{}/frameworks/{}'.format(self.supplier_id, self.framework_slug),
+            data=json.dumps(self.updater_json),
+            content_type='application/json')
+        assert response.status_code == 201
+
+    def _put_variation_agreement(self):
+        data = {"agreedVariations": {"agreedUserId": self.users[0].get("id")}}
+        data.update(self.updater_json)
+
+        response = self.client.put(
+            '/suppliers/{}/frameworks/{}/variation/1'.format(self.supplier_id, self.framework_slug),
+            data=json.dumps(data),
+            content_type='application/json')
+
+        assert response.status_code == 200
+
+    def _create_and_sign_framework_agreement(self):
+        response = self.client.post(
+            '/agreements',
+            data=json.dumps(
+                {
+                    'updated_by': 'interested@example.com',
+                    'agreement': {
+                        'supplierId': self.supplier_id,
+                        'frameworkSlug': self.framework_slug
+                    },
+                }),
+            content_type='application/json')
+        agreement_id = json.loads(response.get_data(as_text=True))['agreement']['id']
+        self.client.post(
+            "/agreements/{}/sign".format(agreement_id),
+            data=json.dumps({'updated_by': 'interested@example.com'}),
+            content_type='application/json'
+        )
+
+    def _put_declaration(self, status):
+        data = {'declaration': {'status': status}}
+        data.update(self.updater_json)
+
+        response = self.client.get(
+            '/suppliers/{}/frameworks/{}'.format(self.supplier_id, self.framework_slug)
+        )
+
+        current_declaration = json.loads(response.get_data(as_text=True))['frameworkInterest']['declaration']
+        data['declaration'].update(current_declaration)
+
+        response = self.client.put(
+            '/suppliers/{}/frameworks/{}/declaration'.format(self.supplier_id, self.framework_slug),
+            data=json.dumps(data),
+            content_type='application/json')
+
+        assert response.status_code == (200 if current_declaration else 201)
+
+    def _put_complete_declaration(self):
+        self._put_declaration(status='complete')
+
+    def _put_incomplete_declaration(self):
+        self._put_declaration(status='started')
+
+    def _get_declaration(self):
+        response = self.client.get('/suppliers/{}/frameworks/{}'.format(self.supplier_id, self.framework_slug))
+        return json.loads(response.get_data(as_text=True))['frameworkInterest']['declaration']
+
+    def _post_company_details_confirmed(self):
+        response = self.client.post(
+            f'/suppliers/{self.supplier_id}',
+            data=json.dumps({
+                "updated_by": "Miss Fig",
+                "suppliers": {
+                    "companyDetailsConfirmed": True
+                }
+            }),
+            content_type='application/json',
+        )
+
+        assert response.status_code == 200
+
+        # Company details must be confirmed at the supplier account level as well as the application level.
+        response = self.client.post(
+            '/suppliers/{}/frameworks/{}'.format(self.supplier_id, self.framework_slug),
+            data=json.dumps({
+                "updated_by": "Mr Sausages",
+                'frameworkInterest': {
+                    'applicationCompanyDetailsConfirmed': True
+                }
+            }),
+            content_type='application/json')
+
+        assert response.status_code == 200
+
+    def _post_complete_draft_service(self):
+        payload = load_example_listing("DOS-digital-specialist")
+
+        self.draft_json = {'services': payload}
+        self.draft_json['services']['supplierId'] = self.supplier_id
+        self.draft_json['services']['frameworkSlug'] = self.framework_slug
+        self.draft_json.update(self.updater_json)
+
+        response = self.client.post(
+            '/draft-services',
+            data=json.dumps(self.draft_json),
+            content_type='application/json')
+
+        assert response.status_code == 201
+
+        draft_id = json.loads(response.get_data())['services']['id']
+        complete = self.client.post(
+            '/draft-services/{}/complete'.format(draft_id),
+            data=json.dumps(self.updater_json),
+            content_type='application/json')
+
+        assert complete.status_code == 200
+
+
 def load_example_listing(name):
     file_path = os.path.join('example_listings', '{}.json'.format(name))
     with open(file_path) as f:
